+++ lib/lennarb.rb
@@ -1,14 +1,15 @@
 # frozen_string_literal: true
 
-# Set environment
-#
-ENV['RACK_ENV'] ||= 'development'
+# Released under the MIT License.
+# Copyright, 2023, by Aristóteles Coutinho.
+
+ENV["RACK_ENV"] ||= "development"
 
 # Extension for Array class
 #
-require 'lennarb/array_extensions'
+require "lennarb/array_extensions"
 
 # Base class for Lennarb
 #
-require 'lenna/application'
-require 'lennarb/version'
+require "lenna/application"
+require "lennarb/version"