+++ test/lib/lenna/middleware/test_app.rb
@@ -1,6 +1,9 @@
 # frozen_string_literal: true
 
-require 'test_helper'
+# Released under the MIT License.
+# Copyright, 2023, by Aristóteles Coutinho.
+
+require "test_helper"
 
 module Lenna
   module Middleware
@@ -12,7 +15,7 @@ module Lenna
       def test_reset
         mdw =
           lambda do |_req, res, next_middleware|
-            res.assign_header('Custom-Header', 'From mdw')
+            res.assign_header("Custom-Header", "From mdw")
             next_middleware.call
           end
 
@@ -24,7 +27,7 @@ module Lenna
       def test_use
         mdw =
           lambda do |_req, res, next_middleware|
-            res.assign_header('Custom-Header', 'From mdw')
+            res.assign_header("Custom-Header", "From mdw")
             next_middleware.call
           end
 
@@ -41,29 +44,29 @@ module Lenna
       end
 
       def test_middleware_execution_order_and_final_state
-        env      = ::Rack::MockRequest.env_for('/')
-        request  = Lenna::Router::Request.new(env)
+        env = ::Rack::MockRequest.env_for("/")
+        request = Lenna::Router::Request.new(env)
         response = Lenna::Router::Response.new
 
         mdw1 =
           lambda do |_req, res, next_middleware|
-            res.assign_header('Custom-Header', 'From mdw1')
+            res.assign_header("Custom-Header", "From mdw1")
             next_middleware.call
           end
         mdw2 =
           lambda do |_req, res, next_middleware|
-            res.assign_header('Custom-Header', 'From mdw2')
+            res.assign_header("Custom-Header", "From mdw2")
             next_middleware.call
           end
 
         @app.use([mdw1, mdw2])
-        action = ->(_req, res) { [200, res.headers, ['OK']] }
+        action = ->(_req, res) { [200, res.headers, ["OK"]] }
         chain = @app.build_middleware_chain(action, [])
         status, headers, body = chain.call(request, response)
 
         assert_equal(200, status)
-        assert_equal({ 'Custom-Header' => 'From mdw1, From mdw2' }, headers)
-        assert_equal(['OK'], body)
+        assert_equal({"Custom-Header" => "From mdw1, From mdw2"}, headers)
+        assert_equal(["OK"], body)
       end
     end
   end